Converted the whole game from int system to float system. Added 'SmartEnemy'.

+++ src/com/tutorial/game/BasicEnemy.java
@@ -17,7 +17,7 @@ public class BasicEnemy extends GameObject {
 	}
 	
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 	
 	public void tick() {
@@ -31,12 +31,12 @@ public class BasicEnemy extends GameObject {
 		if (x <= 0 || x >= Game.WIDTH - 20) vX *= -1;
 		
 		// Add trail object
-		handler.addObject(new Trail(x, y, 16, 16, Color.red, (float) 0.04, ID.Trail, handler));
+		handler.addObject(new Trail((int)x, (int)y, 16, 16, Color.red, (float) 0.04, ID.Trail, handler));
 	}
 	
 	public void render(Graphics g) {
 		g.setColor(Color.red);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 	
 }
